+++ my-project/package.json
@@ -12,6 +12,8 @@
     "lint-fix": "npx eslint . --fix"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.0.12",
+    "@react-navigation/native-stack": "^6.8.0",
     "expo": "~44.0.0",
     "expo-location": "~14.0.1",
     "expo-status-bar": "~1.2.0",
@@ -20,7 +22,9 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-maps": "^1.3.1",
-    "react-native-web": "0.17.1"
+    "react-native-web": "0.17.1",
+    "react-native-screens": "~3.10.1",
+    "react-native-safe-area-context": "3.3.2"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
